requirements.txt
anyio==4.11.0
asgiref==3.10.0
blinker==1.9.0
certifi==2025.10.5
click==8.3.0
Django==5.2.7
Flask==3.1.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
tenacity==9.1.2
typing_extensions==4.15.0
Werkzeug==3.1.3
wheel==0.45.1
